Fix mistargetting the wrong librhonabwy.so UUT
Commit 9514d739 broke the test build script by causing the test suite
to be built with LDFLAGS="-L$(RHONABWY_LIBRARY) -lrhonabwy" instead of
LDFLAGS="-L$(RHONABWY_LOCATION) -lrhonabwy"

This caused test to be linked against the /usr/lib version of
librhonabwy.so instead of the ../src version.

One effect is that the test suite could not be built until the library
was first installed.

This commit enables linking the test suite to be correct again.